morningfocus/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morningfocus"
version = "0.1.0"
description = "Local CLI task manager with morning briefings"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
notify = ["plyer>=2.1"]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "typer[all]>=0.12",
]

[project.scripts]
mf = "morningfocus.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["morningfocus*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["morningfocus"]
omit = ["tests/*", "notify.py"]
